Changed modifiers to double Added return values to interface Moved CharacterLevel and CharacterPoints from Game to Character

+++ src/factory/race/DarkElf.java
@@ -1,7 +1,7 @@
 package factory.race;
 
 public class DarkElf extends Race {
-    public DarkElf(int strength, int intelligence, int agility, int toughness) {
+    public DarkElf(double strength, double intelligence, double agility, double toughness) {
         super(strength, intelligence, agility, toughness);
     }
 }
