+++ src/factory/race/Dwarf.java
@@ -1,7 +1,7 @@
 package factory.race;
 
 public class Dwarf extends Race {
-    public Dwarf(int strength, int intelligence, int agility, int toughness) {
+    public Dwarf(double strength, double intelligence, double agility, double toughness) {
         super(strength, intelligence, agility, toughness);
     }
 }
